refacto: Rename onGasStation to GasStation
