+++ src/app/gasStation/components/Gauge/Electric.js
