+++ src/app/gasStation/components/Gauge/Fuel.js
