+++ src/app/gasStation/components/Gauge/index.js
