+++ src/app/gasStation/components/Gauge/styles.js
