+++ src/app/gasStation/components/Payment/styles.js
