+++ src/app/gasStation/redux/actions/player.js
