+++ src/app/gasStation/redux/reducers/entities/gases/getters/get-gas-is-in-stock.js
