+++ src/app/gasStation/redux/reducers/entities/player/getters/get-has-no-enough-money.js
