+++ src/app/gasStation/redux/reducers/entities/player/getters/get-remaining-money-from-player.js
