+++ src/app/gasStation/redux/reducers/entities/player/index.js
