+++ src/app/gasStation/redux/reducers/entities/player/selectors/select-fuel-value-from-car.js
