+++ src/app/gasStation/redux/reducers/entities/player/selectors/select-money-from-player.js
