+++ src/app/gasStation/redux/reducers/entities/player/selectors/select-tank-from-car.js
