+++ src/app/gasStation/redux/reducers/entities/selectors/select-player.js
