+++ src/app/gasStation/redux/reducers/entities/station/index.js
