+++ src/app/gasStation/redux/reducers/entities/station/selectors/select-station.js
