+++ src/app/gasStation/redux/reducers/index.js
