+++ src/app/gasStation/redux/reducers/sceneState/getters/get-current-gas-selected.js
