+++ src/app/gasStation/redux/reducers/sceneState/getters/get-has-buy-disabled.js
