+++ src/app/gasStation/redux/reducers/sceneState/getters/get-has-maximum-quantity-reached.js
