+++ src/app/gasStation/redux/reducers/sceneState/index.js
