+++ src/app/gasStation/redux/reducers/sceneState/selectors/select-count-buy.js
