+++ src/app/gasStation/redux/reducers/sceneState/selectors/select-gas-selected-id.js
