+++ src/app/gasStation/redux/reducers/sceneState/selectors/select-scene-state.js
