+++ src/app/gasStation/redux/reducers/sceneState/selectors/select-shopping-cart.js
