+++ src/app/gasStation/redux/sagas/fixtures/gasStation.js
