+++ src/app/gasStation/redux/sagas/player.js
