+++ src/gasStation.html
@@ -8,7 +8,7 @@
     <meta name="description" content="">
 </head>
 <body>
-<div id="onGasStation"></div>
+<div id="BURDGIALAX_GasStation"></div>
 </body>
 <script>
     (function(obj)
