+++ src/onShop.html
@@ -8,7 +8,7 @@
     <meta name="description" content="">
 </head>
 <body>
-<div id="onShop"></div>
+<div id="BURDIGALAX_onShop"></div>
 </body>
 <script>
     (function(obj)
