+++ src/onShop.js
@@ -3,6 +3,6 @@ import ReactDOM from "react-dom";
 import OnShop from "./app/onShop";
 import "./Montserrat.css";
 
-var mountNode = document.getElementById("onShop");
+var mountNode = document.getElementById("BURDIGALAX_onShop");
 
 ReactDOM.render(<OnShop />, mountNode);
